lxd: split container classes into different files
